Fix Anuncios no requiere props
- En lugar de pasar los props en videowall.jsx, se tienen directamente dentro de anuncios (prevenimos hacer muchas consultas al actualizar el componente principal)
- Tal vez valga la pena antes de declarar activeSlide hacer la consulta a la bd (usa data.length)
