+++ DREAM Lab Frontend/src/pages/Videowall/components/Anuncios/Anuncios.jsx
@@ -1,22 +1,47 @@
 import React, { useState, useEffect} from "react";
 import "./Anuncios.css";
 
-function Anuncios(props) {
+const anunciosDefault = [
+    {
+        id: 1,
+        img: "https://media.licdn.com/dms/image/C4E12AQHTZ7pyTmCfEw/article-cover_image-shrink_720_1280/0/1541423067327?e=2147483647&v=beta&t=4fBTi39dOwxthLmtYW4RGwseAA9kAXBPUdnQa_7PKx4",
+        nombre: "Lab Default 1",
+        sala: "Sala Default 1",
+        date: "Lunes 15 de Febrero 3:00pm a 5:00pm",
+    },
+    {
+        id: 2,
+        img: "https://mms.businesswire.com/media/20190617005571/es/727944/5/Warner_Final_Zone5.1.jpg",
+        nombre: "Lab Default 2",
+        sala: "Sala Default 2",
+        date: "Lunes 15 de Febrero 3:00pm a 5:00pm",
+    },
+    {
+        id: 3,
+        img: "https://d1tm14lrsghf7q.cloudfront.net/public/media/23459/conversions/21386-thumb.jpg",
+        nombre: "Lab Default 3",
+        sala: "Sala Default 3",
+        date: "Lunes 15 de Febrero 3:00pm a 5:00pm",
+    },
+];
+
+function Anuncios() {
+    const [data, setData] = useState(anunciosDefault);
     // Estados para manejar el anuncio mostrado y que rote automáticamente
-    const [activeSlide, setActiveSlide] = useState(props.activeSlide);
+    const [activeSlide, setActiveSlide] = useState(parseInt(Math.floor(data.length / 2)));
     const [autoRotate, setAutoRotate] = useState(true);
 
-    console.log(props.data);
+    //console.log("Data: ", data);
 
 
     // Funciones para rotar el carrusel
     const next = () => {
-      setActiveSlide((activeSlide + 1) % props.data.length);
+      setActiveSlide((activeSlide + 1) % data.length);
     };
 
     const prev = () => {
         setActiveSlide(
-            activeSlide === 0 ? props.data.length - 1 : activeSlide - 1
+            activeSlide === 0 ? data.length - 1 : activeSlide - 1
         );
     };
     
@@ -49,7 +74,7 @@ function Anuncios(props) {
 
     return (
       <div className="containerAnuncios">
-                {props.data.map((item, i) => (
+                {data.map((item, i) => (
                     <React.Fragment key={i}>
                         <div
                             className="slideAnuncio"
